services/ai-detection/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
numpy==1.26.3
pydantic==2.5.3

# New additions for enhanced detection
opencv-python==4.9.0.80
scikit-image==0.22.0
imagehash==4.3.1
exifread==3.0.0
scipy==1.11.4
